requirements.txt
Babel==2.5.1
certifi==2017.7.27.1
chardet==3.0.4
codegen==1.0
Django==1.10.8
django-ckeditor==5.3.0
django-jet==1.0.6
django-js-asset==0.1.1
django-modeltranslation==0.12.1
django-phonenumber-field==1.3.0
django-rosetta==0.7.13
django-widget-tweaks==1.4.1
idna==2.6
microsofttranslator==0.8
MySQL-python==1.2.5
olefile==0.44
patterns==0.3
phonenumberslite==8.8.2
Pillow==4.2.1
pluggy==0.5.2
polib==1.0.8
py==1.4.34
pytz==2017.2
redis==2.10.6
requests==2.18.4
simplejson==3.11.1
six==1.11.0
solid-i18n==1.4.1
sorl-thumbnail==12.3
tox==2.9.0
urllib3==1.22
virtualenv==15.1.0
